requirements.txt
anyio==4.10.0
atlassian-python-api==4.0.7
beautifulsoup4==4.13.5
bitbucket-python==0.3.2
certifi==2025.8.3
charset-normalizer==3.4.3
defusedxml==0.7.1
deprecated==1.2.18
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
jira==3.10.5
jmespath==1.0.1
oauthlib==3.3.1
packaging==25.0
python-dotenv==1.1.1
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
sniffio==1.3.1
soupsieve==2.8
typing-extensions==4.15.0
urllib3==2.5.0
wrapt==1.17.3
